packages/app-config/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "app-config"
version = "1.0.0"
description = "Application configuration management for IBKR Portfolio Rebalancer"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
